refactor: move use-cases to src
